requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pkcs7==0.1.2
pycryptodome==3.19.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
urllib3==2.0.7
